backend/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.2.1
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
webcolors==1.11.1
asgiref==3.2.10
django-filter==2.4.0
gunicorn==20.0.4
django-environ==0.8.0
flake8==4.0.1
flake8-quotes==3.3.